Remove unused composables
