+++ app/src/main/java/com/babylon/wallet/android/presentation/account/composable/LSUBottomSheetDetails.kt
@@ -29,12 +29,12 @@ import com.babylon.wallet.android.domain.model.resources.XrdResource
 import com.babylon.wallet.android.presentation.ui.composables.BackIconType
 import com.babylon.wallet.android.presentation.ui.composables.RadixCenteredTopAppBar
 import com.babylon.wallet.android.presentation.ui.composables.Thumbnail
+import com.babylon.wallet.android.presentation.ui.composables.assets.poolName
 import com.babylon.wallet.android.presentation.ui.composables.icon
 import com.babylon.wallet.android.presentation.ui.composables.name
 import com.babylon.wallet.android.presentation.ui.composables.resources.AddressRow
 import com.babylon.wallet.android.presentation.ui.composables.resources.TokenBalance
 import com.babylon.wallet.android.presentation.ui.composables.resources.ValidatorDetailsItem
-import com.babylon.wallet.android.presentation.ui.composables.resources.poolName
 import rdx.works.core.displayableQuantity
 import java.math.BigDecimal
