+++ app/src/main/java/com/babylon/wallet/android/presentation/account/composable/PoolUnitBottomSheetDetails.kt
@@ -25,12 +25,12 @@ import com.babylon.wallet.android.domain.model.assets.PoolUnit
 import com.babylon.wallet.android.presentation.ui.composables.BackIconType
 import com.babylon.wallet.android.presentation.ui.composables.RadixCenteredTopAppBar
 import com.babylon.wallet.android.presentation.ui.composables.Thumbnail
+import com.babylon.wallet.android.presentation.ui.composables.assets.PoolResourcesValues
+import com.babylon.wallet.android.presentation.ui.composables.assets.poolName
 import com.babylon.wallet.android.presentation.ui.composables.icon
 import com.babylon.wallet.android.presentation.ui.composables.name
 import com.babylon.wallet.android.presentation.ui.composables.resources.AddressRow
-import com.babylon.wallet.android.presentation.ui.composables.resources.PoolResourcesValues
 import com.babylon.wallet.android.presentation.ui.composables.resources.TokenBalance
-import com.babylon.wallet.android.presentation.ui.composables.resources.poolName
 
 @Composable
 fun PoolUnitBottomSheetDetails(
