--- app/src/main/java/com/babylon/wallet/android/presentation/ui/composables/resources/FungibleResourceItem.kt
@@ -1,110 +1,0 @@
-package com.babylon.wallet.android.presentation.ui.composables.resources
-
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.layout.width
-import androidx.compose.material.Text
-import androidx.compose.material3.Checkbox
-import androidx.compose.material3.CheckboxDefaults
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.unit.dp
-import com.babylon.wallet.android.designsystem.theme.RadixTheme
-import com.babylon.wallet.android.domain.model.resources.Resource
-import com.babylon.wallet.android.presentation.ui.composables.Thumbnail
-import rdx.works.core.displayableQuantity
-
-@Composable
-fun FungibleResourceItem(
-    resource: Resource.FungibleResource,
-    modifier: Modifier = Modifier,
-    trailingContent: @Composable (() -> Unit) = {
-        Spacer(modifier = Modifier.width(28.dp))
-    },
-    bottomContent: @Composable (() -> Unit) = {}
-) {
-    Column(modifier = modifier, horizontalAlignment = Alignment.CenterHorizontally) {
-        Row(
-            verticalAlignment = Alignment.CenterVertically,
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(
-                    horizontal = RadixTheme.dimensions.paddingDefault,
-                    vertical = RadixTheme.dimensions.paddingLarge
-                ),
-        ) {
-            Spacer(modifier = Modifier.width(RadixTheme.dimensions.paddingMedium))
-            Thumbnail.Fungible(
-                modifier = Modifier.size(44.dp),
-                token = resource
-            )
-            Spacer(modifier = Modifier.width(RadixTheme.dimensions.paddingMedium))
-            Text(
-                modifier = Modifier.weight(1f),
-                text = resource.displayTitle,
-                style = RadixTheme.typography.body2HighImportance,
-                color = RadixTheme.colors.gray1,
-                maxLines = 1
-            )
-            Spacer(modifier = Modifier.width(RadixTheme.dimensions.paddingMedium))
-
-            resource.ownedAmount?.let { amount ->
-                Text(
-                    text = amount.displayableQuantity(),
-                    style = RadixTheme.typography.secondaryHeader,
-                    color = RadixTheme.colors.gray1,
-                    maxLines = 2
-                )
-            }
-
-            trailingContent()
-        }
-        bottomContent()
-    }
-}
-
-@Composable
-fun SelectableFungibleResourceItem(
-    modifier: Modifier = Modifier,
-    resource: Resource.FungibleResource,
-    isSelected: Boolean,
-    onCheckChanged: (Boolean) -> Unit
-) {
-    FungibleResourceItem(
-        modifier = modifier,
-        resource = resource,
-        trailingContent = {
-            Checkbox(
-                checked = isSelected,
-                onCheckedChange = onCheckChanged,
-                colors = CheckboxDefaults.colors(
-                    checkedColor = RadixTheme.colors.gray1,
-                    uncheckedColor = RadixTheme.colors.gray2,
-                    checkmarkColor = Color.White
-                )
-            )
-
-            Spacer(modifier = Modifier.width(RadixTheme.dimensions.paddingSmall))
-        }
-//        bottomContent = {
-//            resource.tags.find { it == Resource.Tag.Official }?.let { _ ->
-//                Tag(
-//                    modifier = Modifier
-//                        .fillMaxWidth()
-//                        .background(RadixTheme.colors.gray5, shape = RadixTheme.shapes.roundedRectBottomMedium)
-//                        .padding(
-//                            horizontal = RadixTheme.dimensions.paddingMedium,
-//                            vertical = RadixTheme.dimensions.paddingSmall
-//                        ),
-//                    tag = officialTag
-//                )
-//            }
-//        }
-    )
-}
