--- app/src/main/java/com/babylon/wallet/android/presentation/ui/composables/resources/NonFungibleResourceCard.kt
@@ -1,255 +1,0 @@
-package com.babylon.wallet.android.presentation.ui.composables.resources
-
-import androidx.compose.foundation.clickable
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.requiredHeight
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material.Card
-import androidx.compose.material.ExperimentalMaterialApi
-import androidx.compose.material.Surface
-import androidx.compose.material.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.scale
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.Dp
-import androidx.compose.ui.unit.dp
-import com.babylon.wallet.android.R
-import com.babylon.wallet.android.designsystem.theme.RadixTheme
-import com.babylon.wallet.android.designsystem.theme.RadixWalletTheme
-import com.babylon.wallet.android.domain.model.resources.Resource
-import com.babylon.wallet.android.domain.model.resources.metadata.NameMetadataItem
-import com.babylon.wallet.android.presentation.ui.composables.Thumbnail
-
-@OptIn(ExperimentalMaterialApi::class)
-@Composable
-fun NonFungibleResourceCollectionHeader(
-    modifier: Modifier = Modifier,
-    collection: Resource.NonFungibleResource,
-    cardHeight: Dp = 103.dp,
-    collapsed: Boolean = false,
-    groupInnerPadding: Dp = 6.dp,
-    parentSectionClick: () -> Unit,
-) {
-    val bottomCorners = if (collapsed) 12.dp else 0.dp
-    val cardShape = RoundedCornerShape(12.dp, 12.dp, bottomCorners, bottomCorners)
-    Box(
-        modifier = modifier
-    ) {
-        if (collapsed) {
-            if (collection.amount >= 2) {
-                val scaleFactor = 0.8f
-                val topOffset = cardHeight * (1 - scaleFactor) + groupInnerPadding + groupInnerPadding * scaleFactor
-                Surface(
-                    modifier = Modifier
-                        .padding(top = topOffset)
-                        .fillMaxWidth()
-                        .height(cardHeight)
-                        .scale(scaleFactor, scaleFactor),
-                    shape = RadixTheme.shapes.roundedRectMedium,
-                    color = Color.White,
-                    elevation = 2.dp,
-                    content = {}
-                )
-            }
-
-            if (collection.amount > 0) {
-                val scaleFactor = 0.9f
-                val topOffset = cardHeight * (1 - scaleFactor) + groupInnerPadding
-                Surface(
-                    modifier = Modifier
-                        .padding(top = topOffset)
-                        .fillMaxWidth()
-                        .height(cardHeight)
-                        .scale(scaleFactor, scaleFactor),
-                    shape = RadixTheme.shapes.roundedRectMedium,
-                    color = Color.White,
-                    elevation = 3.dp,
-                    content = {}
-                )
-            }
-        }
-
-        Card(
-            modifier = Modifier
-                .fillMaxWidth()
-                .requiredHeight(cardHeight)
-                .clickable { parentSectionClick() },
-            shape = cardShape,
-            backgroundColor = Color.White,
-            elevation = 4.dp,
-            onClick = parentSectionClick
-        ) {
-            Row(
-                verticalAlignment = Alignment.CenterVertically,
-                modifier = Modifier
-                    .padding(horizontal = RadixTheme.dimensions.paddingLarge),
-                horizontalArrangement = Arrangement.spacedBy(RadixTheme.dimensions.paddingDefault)
-            ) {
-                Thumbnail.NonFungible(
-                    modifier = Modifier.size(44.dp),
-                    collection = collection,
-                    shape = RadixTheme.shapes.roundedRectSmall
-                )
-                Column(verticalArrangement = Arrangement.Center) {
-                    if (collection.name.isNotEmpty()) {
-                        Text(
-                            collection.name,
-                            style = RadixTheme.typography.secondaryHeader,
-                            color = RadixTheme.colors.gray1,
-                            maxLines = 2
-                        )
-                    }
-
-                    collection.currentSupply?.let { currentSupply ->
-                        Text(
-                            stringResource(
-                                id = R.string.assetDetails_NFTDetails_ownedOfTotal,
-                                collection.items.size,
-                                currentSupply
-                            ),
-                            style = RadixTheme.typography.body2HighImportance,
-                            color = RadixTheme.colors.gray2,
-                        )
-                    }
-                }
-            }
-        }
-    }
-}
-
-@Preview(showBackground = true)
-@Composable
-fun CollapsableParentItemPreview() {
-    RadixWalletTheme {
-        NonFungibleResourceCollectionHeader(
-            modifier = Modifier.padding(all = 30.dp),
-            collection = Resource.NonFungibleResource(
-                resourceAddress = "resource_rdx_abcde",
-                amount = 1,
-                nameMetadataItem = NameMetadataItem(name = "Crypto Punks"),
-                items = listOf(
-                    Resource.NonFungibleResource.Item(
-                        collectionAddress = "resource_rdx_abcde",
-                        localId = Resource.NonFungibleResource.Item.ID.from("#1#"),
-                        iconMetadataItem = null,
-                        nameMetadataItem = null
-                    ),
-                    Resource.NonFungibleResource.Item(
-                        collectionAddress = "resource_rdx_abcde",
-                        localId = Resource.NonFungibleResource.Item.ID.from("#2#"),
-                        iconMetadataItem = null,
-                        nameMetadataItem = null
-                    ),
-                    Resource.NonFungibleResource.Item(
-                        collectionAddress = "resource_rdx_abcde",
-                        localId = Resource.NonFungibleResource.Item.ID.from("#3#"),
-                        iconMetadataItem = null,
-                        nameMetadataItem = null
-                    )
-                )
-            ),
-            collapsed = false
-        ) { }
-    }
-}
-
-@Preview(showBackground = true)
-@Composable
-fun ExpandedParentItemPreview() {
-    RadixWalletTheme {
-        NonFungibleResourceCollectionHeader(
-            modifier = Modifier.padding(all = 16.dp),
-            collection = Resource.NonFungibleResource(
-                resourceAddress = "resource_rdx_abcde",
-                amount = 1,
-                nameMetadataItem = NameMetadataItem(name = "Crypto Punks"),
-                items = listOf(
-                    Resource.NonFungibleResource.Item(
-                        collectionAddress = "resource_rdx_abcde",
-                        localId = Resource.NonFungibleResource.Item.ID.from("#1#"),
-                        nameMetadataItem = null,
-                        iconMetadataItem = null
-                    ),
-                    Resource.NonFungibleResource.Item(
-                        collectionAddress = "resource_rdx_abcde",
-                        localId = Resource.NonFungibleResource.Item.ID.from("#2#"),
-                        nameMetadataItem = null,
-                        iconMetadataItem = null
-                    ),
-                    Resource.NonFungibleResource.Item(
-                        collectionAddress = "resource_rdx_abcde",
-                        localId = Resource.NonFungibleResource.Item.ID.from("#3#"),
-                        nameMetadataItem = null,
-                        iconMetadataItem = null
-                    )
-                )
-            ),
-            collapsed = true
-        ) { }
-    }
-}
-
-@Preview(showBackground = true)
-@Composable
-fun ExpandedParentItemPreviewWithTwo() {
-    RadixWalletTheme {
-        NonFungibleResourceCollectionHeader(
-            modifier = Modifier.padding(all = 16.dp),
-            collection = Resource.NonFungibleResource(
-                resourceAddress = "resource_rdx_abcde",
-                amount = 1,
-                nameMetadataItem = NameMetadataItem(name = "Crypto Punks"),
-                items = listOf(
-                    Resource.NonFungibleResource.Item(
-                        collectionAddress = "resource_rdx_abcde",
-                        localId = Resource.NonFungibleResource.Item.ID.from("#1#"),
-                        nameMetadataItem = null,
-                        iconMetadataItem = null
-                    ),
-                    Resource.NonFungibleResource.Item(
-                        collectionAddress = "resource_rdx_abcde",
-                        localId = Resource.NonFungibleResource.Item.ID.from("#2#"),
-                        nameMetadataItem = null,
-                        iconMetadataItem = null
-                    )
-                )
-            ),
-            collapsed = true
-        ) { }
-    }
-}
-
-@Preview(fontScale = 2f, showBackground = true)
-@Composable
-fun CollapsableParentItemWithLargeFontPreview() {
-    RadixWalletTheme {
-        NonFungibleResourceCollectionHeader(
-            modifier = Modifier.padding(all = 16.dp),
-            collection = Resource.NonFungibleResource(
-                resourceAddress = "resource_rdx_abcde",
-                amount = 1,
-                nameMetadataItem = NameMetadataItem(name = "Crypto Punks"),
-                items = listOf(
-                    Resource.NonFungibleResource.Item(
-                        collectionAddress = "resource_rdx_abcde",
-                        localId = Resource.NonFungibleResource.Item.ID.from("#1#"),
-                        nameMetadataItem = null,
-                        iconMetadataItem = null
-                    )
-                )
-            ),
-            collapsed = false
-        ) { }
-    }
-}
